requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.1.0
numpy>=1.25.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
tqdm>=4.66.1
python-dotenv>=1.0.0

# PDF and document parsing
pymupdf>=1.23.0
pdfminer.six>=20221105

# Data formats and I/O
pyarrow>=14.0.1
fastparquet>=2023.8.0

# Google Cloud
google-cloud-storage>=2.10.0

# Embeddings & Vector Store
weaviate-client>=4.4.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Optional: OpenAI fallback
openai>=1.0.0

# Concurrency and async utils
aiohttp>=3.9.0
tenacity>=8.2.3
urllib3>=2.0.7
idna>=3.4

# Headless browser fallback (for blocked PDF access)
playwright>=1.47.0

# Data ingestion
grobid-client-python

# Core Testing Framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0

# Data Quality & Validation
great-expectations==0.18.8
pandera==0.17.2

# Mocking & Fixtures
responses==0.24.1
faker==20.1.0
freezegun==1.4.0

# Code Quality
pylint==3.0.3
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Optional but useful
pytest-html==4.1.1
pytest-json-report==1.5.0

dvc[gs]==3.32.0

# Orchestration (optional for now)
# apache-airflow>=2.7.0

# Neo4j (for future citation graph)
# neo4j>=5.0.0

# PostgreSQL (for future metadata store)
# psycopg2-binary>=2.9.0

# ------------------------------
# CiteConnect Data Bias Analysis
# ------------------------------

# Core data handling
pandas>=2.0.0
numpy>=1.24.0

# GCS integration
google-cloud-storage>=2.10.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.12.2

# Fairness & Slicing
fairlearn>=0.13.0
scikit-learn>=1.3.0

# Email testing (optional for local runs)
python-dotenv>=1.0.1

# For JSON and file handling
pyarrow>=15.0.0
